refactor(birdie): move config endpoints into lfs
